nfl_ev_betting_engine/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
nfl_data_py>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
joblib>=1.3.0
appdirs>=1.4.0
fastparquet>=2024.0.0
pytest>=8.0.0
